Update SrtWriter.hx
This fixes some issues with the .srt writer file. it fixes the end time because in the @SBT the "end time" is actually just a duration of the subtitle but .srt files define the end time as a timestamp, not a duration so this adds the end time duration to the start time to get a correct end time before writing it to the .srt.

it also fixes an issue where anything over 100 seconds gets messed up due to the .srt file going from 0:99 start time to 1:00 start time which is 1 minute and is 40 seconds behind the 100 second actual start time of the timestamp. this accounts for adding 40 seconds every 100 seconds to the start time and accounts for .usms up to 20 minutes long